Optimize Custom Crafting system into Custom Classes
