--- mcp/src/minecraft/assets/minecraft/models/item/custom_crafting_table.json
@@ -1,3 +1,0 @@
-{
-    "parent": "block/custom_crafting_table"
-}
