--- mcp/src/minecraft/com/elementfx/tvp/ad/client/gui/inventory/GuiGoblinCrafting.java
@@ -1,46 +1,0 @@
-package com.elementfx.tvp.ad.client.gui.inventory;
-import com.elementfx.tvp.ad.inventory.ContainerGoblinWorkbench;
-import net.minecraft.client.gui.inventory.GuiContainer;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.resources.I18n;
-import net.minecraft.entity.player.InventoryPlayer;
-import net.minecraft.inventory.ContainerWorkbench;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.World;
-
-public class GuiGoblinCrafting extends GuiContainer
-{
-	private static final ResourceLocation GOBLIN_CRAFTING_TABLE_GUI_TEXTURES = new ResourceLocation("textures/gui/container/crafting_table.png");
-	
-    public GuiGoblinCrafting(InventoryPlayer playerInv, World worldIn)
-    {
-        this(playerInv, worldIn, BlockPos.ORIGIN);
-    }
-
-    public GuiGoblinCrafting(InventoryPlayer playerInv, World worldIn, BlockPos blockPosition)
-    {
-        super(new ContainerGoblinWorkbench(playerInv, worldIn, blockPosition));
-    }
-
-    /**
-     * Draw the foreground layer for the GuiContainer (everything in front of the items)
-     */
-    protected void drawGuiContainerForegroundLayer(int mouseX, int mouseY)
-    {
-        this.fontRendererObj.drawString(I18n.format("container.goblinCrafting", new Object[0]), 28, 6, 4210752);
-        this.fontRendererObj.drawString(I18n.format("container.inventory", new Object[0]), 8, this.ySize - 96 + 2, 4210752);
-    }
-
-    /**
-     * Draws the background layer of this container (behind the items).
-     */
-    protected void drawGuiContainerBackgroundLayer(float partialTicks, int mouseX, int mouseY)
-    {
-        GlStateManager.color(1.0F, 1.0F, 1.0F, 1.0F);
-        this.mc.getTextureManager().bindTexture(GOBLIN_CRAFTING_TABLE_GUI_TEXTURES);
-        int i = (this.width - this.xSize) / 2;
-        int j = (this.height - this.ySize) / 2;
-        this.drawTexturedModalRect(i, j, 0, 0, this.xSize, this.ySize);
-    }
-}
