--- mcp/src/minecraft/com/elementfx/tvp/ad/inventory/SlotGoblinCrafting.java
@@ -1,63 +1,0 @@
-package com.elementfx.tvp.ad.inventory;
-
-import javax.annotation.Nullable;
-
-import com.elementfx.tvp.ad.item.crafting.ElvenCraftingManager;
-import com.elementfx.tvp.ad.item.crafting.GoblinCraftingManager;
-
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.IInventory;
-import net.minecraft.inventory.InventoryCrafting;
-import net.minecraft.inventory.SlotCrafting;
-import net.minecraft.item.ItemStack;
-
-public class SlotGoblinCrafting extends SlotCrafting
-{
-	/** The craft matrix inventory linked to this result slot. */
-    private final InventoryCrafting craftMatrix;
-
-    /** The player that is using the GUI where this slot resides. */
-    private final EntityPlayer thePlayer;
-
-	public SlotGoblinCrafting(EntityPlayer player, InventoryCrafting craftingInventory, IInventory inventoryIn, int slotIndex, int xPosition, int yPosition) 
-	{
-		super(player, craftingInventory, inventoryIn, slotIndex, xPosition, yPosition);
-		this.thePlayer = player;
-        this.craftMatrix = craftingInventory;
-	}
-	
-	public void onPickupFromSlot(EntityPlayer playerIn, ItemStack stack)
-    {
-        this.onCrafting(stack);
-        ItemStack[] aitemstack = GoblinCraftingManager.getInstance().getRemainingItems(this.craftMatrix, playerIn.worldObj);
-
-        for (int i = 0; i < aitemstack.length; ++i)
-        {
-            ItemStack itemstack = this.craftMatrix.getStackInSlot(i);
-            ItemStack itemstack1 = aitemstack[i];
-
-            if (itemstack != null)
-            {
-                this.craftMatrix.decrStackSize(i, 1);
-                itemstack = this.craftMatrix.getStackInSlot(i);
-            }
-
-            if (itemstack1 != null)
-            {
-                if (itemstack == null)
-                {
-                    this.craftMatrix.setInventorySlotContents(i, itemstack1);
-                }
-                else if (ItemStack.areItemsEqual(itemstack, itemstack1) && ItemStack.areItemStackTagsEqual(itemstack, itemstack1))
-                {
-                    itemstack1.stackSize += itemstack.stackSize;
-                    this.craftMatrix.setInventorySlotContents(i, itemstack1);
-                }
-                else if (!this.thePlayer.inventory.addItemStackToInventory(itemstack1))
-                {
-                    this.thePlayer.dropItem(itemstack1, false);
-                }
-            }
-        }
-    }
-}
