+++ mcp/src/minecraft/com/elementfx/tvp/ad/item/crafting/RohirrimCraftingManager.java
@@ -21,7 +21,7 @@ import net.minecraft.item.crafting.ShapedRecipes;
 import net.minecraft.item.crafting.ShapelessRecipes;
 import net.minecraft.world.World;
 
-public class RohirrimCraftingManager
+public class RohirrimCraftingManager extends CustomCraftingManager
 {
 	 /** The static instance of this class */
     private static final RohirrimCraftingManager INSTANCE = new RohirrimCraftingManager();
@@ -44,155 +44,5 @@ public class RohirrimCraftingManager
                 return p_compare_1_ instanceof ShapelessRecipes && p_compare_2_ instanceof ShapedRecipes ? 1 : (p_compare_2_ instanceof ShapelessRecipes && p_compare_1_ instanceof ShapedRecipes ? -1 : (p_compare_2_.getRecipeSize() < p_compare_1_.getRecipeSize() ? -1 : (p_compare_2_.getRecipeSize() > p_compare_1_.getRecipeSize() ? 1 : 0)));
             }
         });
-    }
-    /**
-     * Adds a shaped recipe to the games recipe list.
-     */
-    public ShapedRecipes addRecipe(ItemStack stack, Object... recipeComponents)
-    {
-        String s = "";
-        int i = 0;
-        int j = 0;
-        int k = 0;
-
-        if (recipeComponents[i] instanceof String[])
-        {
-            String[] astring = (String[])((String[])recipeComponents[i++]);
-
-            for (String s2 : astring)
-            {
-                ++k;
-                j = s2.length();
-                s = s + s2;
-            }
-        }
-        else
-        {
-            while (recipeComponents[i] instanceof String)
-            {
-                String s1 = (String)recipeComponents[i++];
-                ++k;
-                j = s1.length();
-                s = s + s1;
-            }
-        }
-
-        Map<Character, ItemStack> map;
-
-        for (map = Maps.<Character, ItemStack>newHashMap(); i < recipeComponents.length; i += 2)
-        {
-            Character character = (Character)recipeComponents[i];
-            ItemStack itemstack = null;
-
-            if (recipeComponents[i + 1] instanceof Item)
-            {
-                itemstack = new ItemStack((Item)recipeComponents[i + 1]);
-            }
-            else if (recipeComponents[i + 1] instanceof Block)
-            {
-                itemstack = new ItemStack((Block)recipeComponents[i + 1], 1, 32767);
-            }
-            else if (recipeComponents[i + 1] instanceof ItemStack)
-            {
-                itemstack = (ItemStack)recipeComponents[i + 1];
-            }
-
-            map.put(character, itemstack);
-        }
-
-        ItemStack[] aitemstack = new ItemStack[j * k];
-
-        for (int l = 0; l < j * k; ++l)
-        {
-            char c0 = s.charAt(l);
-
-            if (map.containsKey(Character.valueOf(c0)))
-            {
-                aitemstack[l] = ((ItemStack)map.get(Character.valueOf(c0))).copy();
-            }
-            else
-            {
-                aitemstack[l] = null;
-            }
-        }
-
-        ShapedRecipes shapedrecipes = new ShapedRecipes(j, k, aitemstack, stack);
-        this.recipes.add(shapedrecipes);
-        return shapedrecipes;
-    }
-
-    /**
-     * Adds a shapeless crafting recipe to the the game.
-     */
-    public void addShapelessRecipe(ItemStack stack, Object... recipeComponents)
-    {
-        List<ItemStack> list = Lists.<ItemStack>newArrayList();
-
-        for (Object object : recipeComponents)
-        {
-            if (object instanceof ItemStack)
-            {
-                list.add(((ItemStack)object).copy());
-            }
-            else if (object instanceof Item)
-            {
-                list.add(new ItemStack((Item)object));
-            }
-            else
-            {
-                if (!(object instanceof Block))
-                {
-                    throw new IllegalArgumentException("Invalid shapeless recipe: unknown type " + object.getClass().getName() + "!");
-                }
-
-                list.add(new ItemStack((Block)object));
-            }
-        }
-
-        this.recipes.add(new ShapelessRecipes(stack, list));
-    }
-
-    /**
-     * Adds an IRecipe to the list of crafting recipes.
-     */
-    public void addRecipe(IRecipe recipe)
-    {
-        this.recipes.add(recipe);
-    }
-
-    @Nullable
-
-    /**
-     * Retrieves an ItemStack that has multiple recipes for it.
-     */
-    public ItemStack findMatchingRecipe(InventoryCrafting craftMatrix, World worldIn)
-    {
-        for (IRecipe irecipe : this.recipes)
-        {
-            if (irecipe.matches(craftMatrix, worldIn))
-            {
-                return irecipe.getCraftingResult(craftMatrix);
-            }
-        }
-
-        return CraftingManager.getInstance().findMatchingRecipe(craftMatrix, worldIn);
-    }
-
-    public ItemStack[] getRemainingItems(InventoryCrafting craftMatrix, World worldIn)
-    {
-        for (IRecipe irecipe : this.recipes)
-        {
-            if (irecipe.matches(craftMatrix, worldIn))
-            {
-                return irecipe.getRemainingItems(craftMatrix);
-            }
-        }
-
-        return CraftingManager.getInstance().getRemainingItems(craftMatrix, worldIn);
-    }
-
-    public List<IRecipe> getRecipeList()
-    {
-        return this.recipes;
     }
 }
