+++ mcp/src/minecraft/net/minecraft/block/Block.java
@@ -1636,16 +1636,7 @@ public class Block
         registerBlock(783, "diagonal_bricks", (new Block(Material.ROCK)).setHardness(1.5F).setResistance(10.0F).setSoundType(SoundType.STONE).setUnlocalizedName("diagonalBricks").setCreativeTab(CreativeTabs.BUILDING_BLOCKS));
         //registerBlock(784, "dark_brown_wood", (new BlockValarDirectional(Material.wood)).setHardness(2.0F).setSoundType(soundTypeWood).setUnlocalizedName("darkBrownWood").setBlockTextureName("dark_brown_wood"));
         registerBlock(785, "bell", (new BlockBell()).setUnlocalizedName("bell"));
-        Block block500 = (new BlockCustomWorkbench()).setSoundType(SoundType.WOOD).setHardness(0.8F).setUnlocalizedName("customWorkbench");
-        registerBlock(786, "custom_crafting_table", block500);
-        //registerBlock(786, "elven_crafting_table", (new BlockCustomWorkbench()).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("elvenCraftingTable"));
-        //registerBlock(787, "human_crafting_table", (new BlockCustomWorkbench(WorkbenchType.HUMAN)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("humanCraftingTable"));
-        //registerBlock(788, "gondorian_crafting_table", (new BlockCustomWorkbench(WorkbenchType.GONDORIAN)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("gondorianCraftingTable"));
-        //registerBlock(789, "rohirrim_crafting_table", (new BlockCustomWorkbench(WorkbenchType.ROHIRRIM)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("rohirrimCraftingTable"));
-        //registerBlock(790, "hobbit_crafting_table", (new BlockCustomWorkbench(WorkbenchType.HOBBIT)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("hobbitCraftingTable"));
-        //registerBlock(791, "mordor_crafting_table", (new BlockCustomWorkbench(WorkbenchType.MORDOR)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("mordorCraftingTable"));
-        //registerBlock(792, "isengard_crafting_table", (new BlockCustomWorkbench(WorkbenchType.ISENGARD)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("isengardCraftingTable"));
-        //registerBlock(793, "goblin_crafting_table", (new BlockCustomWorkbench(WorkbenchType.GOBLIN)).setHardness(2.5F).setSoundType(SoundType.WOOD).setUnlocalizedName("goblinCraftingTable"));
+        registerBlock(786, "custom_crafting_table", (new BlockCustomWorkbench()).setSoundType(SoundType.WOOD).setHardness(0.8F).setUnlocalizedName("customWorkbench"));
         // End Awaken Dreams code
         REGISTRY.validateKey();
